// src/app/models/Product.ts
export interface Product {
  name: String;
  price: Number;
  description: String;
  image: string;
  url: string;
  coupon: String;
  special_price: Number;
  store_id: any;
  expires_on: String;
  viewed: Number;
  created_by: Number;
  reviewed_by: Number;
  status: Number;
  rating: Number;
  offer_type: string;
}